+++ src/Services/PaymentService.php
@@ -229,6 +229,7 @@ class PaymentService
      *                                            yet (i.e. has no id).
      * @param bool|null              $card3ds     Enables 3ds channel for credit cards if available. This parameter is
      *                                            optional and will be ignored if not applicable.
+     * @param string|null            $invoiceId   The external id of the invoice.
      *
      * @return Charge Resulting Charge object.
      *
@@ -244,10 +245,11 @@ class PaymentService
         $orderId = null,
         $metadata = null,
         $basket = null,
-        $card3ds = null
+        $card3ds = null,
+        $invoiceId = null
     ): AbstractTransactionType {
         $payment = $this->createPayment($paymentType);
-        $charge = (new Charge($amount, $currency, $returnUrl))->setOrderId($orderId);
+        $charge = (new Charge($amount, $currency, $returnUrl))->setOrderId($orderId)->setInvoiceId($invoiceId);
         if ($card3ds !== null) {
             $charge->setCard3ds($card3ds);
         }
@@ -364,15 +366,17 @@ class PaymentService
      *
      * @param Charge $charge
      * @param $amount
+     * @param string|null $reasonCode
      *
      * @return Cancellation Resulting Cancellation object.
      *
      * @throws HeidelpayApiException
      * @throws RuntimeException
      */
-    public function cancelCharge(Charge $charge, $amount = null): AbstractTransactionType
+    public function cancelCharge(Charge $charge, $amount = null, string $reasonCode = null): AbstractTransactionType
     {
         $cancellation = new Cancellation($amount);
+        $cancellation->setReasonCode($reasonCode);
         $cancellation->setPayment($charge->getPayment());
         $charge->addCancellation($cancellation);
         $this->resourceService->create($cancellation);
@@ -388,15 +392,17 @@ class PaymentService
      * Creates a Shipment transaction for the given Payment object.
      *
      * @param Payment|string $payment
+     * @param string|null    $invoiceId
      *
      * @return Shipment Resulting Shipment object.
      *
      * @throws HeidelpayApiException
      * @throws RuntimeException
      */
-    public function ship($payment): AbstractHeidelpayResource
+    public function ship($payment, $invoiceId = null): AbstractHeidelpayResource
     {
         $shipment = new Shipment();
+        $shipment->setInvoiceId($invoiceId);
         $this->resourceService->getPaymentResource($payment)->addShipment($shipment);
         $this->resourceService->create($shipment);
         return $shipment;