+++ test/unit/Resources/TransactionTypes/ChargeTest.php
@@ -52,16 +52,21 @@ class ChargeTest extends BaseUnitTest
         $this->assertNull($charge->getAmount());
         $this->assertNull($charge->getCurrency());
         $this->assertNull($charge->getReturnUrl());
+        $this->assertNull($charge->isCard3ds());
 
         $charge = new Charge(123.4, 'myCurrency', 'https://my-return-url.test');
+        $charge->setCard3ds(true);
         $this->assertEquals(123.4, $charge->getAmount());
         $this->assertEquals('myCurrency', $charge->getCurrency());
         $this->assertEquals('https://my-return-url.test', $charge->getReturnUrl());
+        $this->assertTrue($charge->isCard3ds());
 
         $charge->setAmount(567.8)->setCurrency('myNewCurrency')->setReturnUrl('https://another-return-url.test');
+        $charge->setCard3ds(false);
         $this->assertEquals(567.8, $charge->getAmount());
         $this->assertEquals('myNewCurrency', $charge->getCurrency());
         $this->assertEquals('https://another-return-url.test', $charge->getReturnUrl());
+        $this->assertFalse($charge->isCard3ds());
     }
 
     /**