+++ test/unit/Services/ResourceServiceTest.php
@@ -36,12 +36,14 @@ use heidelpayPHP\Resources\Customer;
 use heidelpayPHP\Resources\Keypair;
 use heidelpayPHP\Resources\Metadata;
 use heidelpayPHP\Resources\Payment;
+use heidelpayPHP\Resources\PaymentTypes\Alipay;
 use heidelpayPHP\Resources\PaymentTypes\BasePaymentType;
 use heidelpayPHP\Resources\PaymentTypes\Card;
 use heidelpayPHP\Resources\PaymentTypes\EPS;
 use heidelpayPHP\Resources\PaymentTypes\Giropay;
 use heidelpayPHP\Resources\PaymentTypes\Ideal;
 use heidelpayPHP\Resources\PaymentTypes\Invoice;
+use heidelpayPHP\Resources\PaymentTypes\InvoiceFactoring;
 use heidelpayPHP\Resources\PaymentTypes\InvoiceGuaranteed;
 use heidelpayPHP\Resources\PaymentTypes\Paypal;
 use heidelpayPHP\Resources\PaymentTypes\PIS;
@@ -50,6 +52,7 @@ use heidelpayPHP\Resources\PaymentTypes\Przelewy24;
 use heidelpayPHP\Resources\PaymentTypes\SepaDirectDebit;
 use heidelpayPHP\Resources\PaymentTypes\SepaDirectDebitGuaranteed;
 use heidelpayPHP\Resources\PaymentTypes\Sofort;
+use heidelpayPHP\Resources\PaymentTypes\Wechatpay;
 use heidelpayPHP\Resources\TransactionTypes\Authorization;
 use heidelpayPHP\Resources\TransactionTypes\Cancellation;
 use heidelpayPHP\Resources\TransactionTypes\Charge;
@@ -1192,6 +1195,30 @@ class ResourceServiceTest extends BaseUnitTest
         $resourceSrvMock->fetchResourceByUrl($resourceUrl);
     }
 
+    /**
+     * Verify does not call fetchResourceByUrl and returns null if the resource type is unknown.
+     *
+     * @test
+     *
+     * @throws Exception
+     * @throws HeidelpayApiException
+     * @throws RuntimeException
+     * @throws ReflectionException
+     * @throws RuntimeException
+     */
+    public function fetchResourceByUrlForAPaymentTypeShouldReturnNullIfTheTypeIsUnknown()
+    {
+        $resourceSrvMock = $this->getMockBuilder(ResourceService::class)->disableOriginalConstructor()
+            ->setMethods(['fetchPaymentType'])->getMock();
+
+        $resourceSrvMock->expects($this->never())->method('fetchPaymentType');
+
+        /** @var ResourceService $resourceSrvMock */
+        $this->assertNull(
+            $resourceSrvMock->fetchResourceByUrl('https://api.heidelpay.com/v1/types/card/s-unknown-xen2ybcovn56/')
+        );
+    }
+
     //<editor-fold desc="Data Providers">
 
     /**
@@ -1260,6 +1287,9 @@ class ResourceServiceTest extends BaseUnitTest
             'Sofort sandbox' => [Sofort::class, 's-sft-12345678'],
             'PIS sandbox' => [PIS::class, 's-pis-12345678'],
             'EPS sandbox' => [EPS::class, 's-eps-12345678'],
+            'Alipay sandbox' => [Alipay::class, 's-ali-12345678'],
+            'Wechatpay sandbox' => [Wechatpay::class, 's-wcp-12345678'],
+            'Invoice factoring sandbox' => [InvoiceFactoring::class, 's-ivf-12345678'],
             'Card production' => [Card::class, 'p-crd-12345678'],
             'Giropay production' => [Giropay::class, 'p-gro-12345678'],
             'Ideal production' => [Ideal::class, 'p-idl-12345678'],
@@ -1271,7 +1301,10 @@ class ResourceServiceTest extends BaseUnitTest
             'SepaDirectDebit production' => [SepaDirectDebit::class, 'p-sdd-12345678'],
             'SepaDirectDebitGuaranteed production' => [SepaDirectDebitGuaranteed::class, 'p-ddg-12345678'],
             'Sofort production' => [Sofort::class, 'p-sft-12345678'],
-            'EPS production' => [EPS::class, 'p-eps-12345678']
+            'EPS production' => [EPS::class, 'p-eps-12345678'],
+            'Alipay production' => [Alipay::class, 'p-ali-12345678'],
+            'Wechatpay production' => [Wechatpay::class, 'p-wcp-12345678'],
+            'Invoice factoring production' => [InvoiceFactoring::class, 'p-ivf-12345678']
         ];
     }
 
@@ -1364,51 +1397,63 @@ class ResourceServiceTest extends BaseUnitTest
             ],
             'GIROPAY'                      => [
                 's-gro-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-gro-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/giropay/s-gro-xen2ybcovn56/'
             ],
             'IDEAL'                        => [
                 's-idl-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-idl-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/ideal/s-idl-xen2ybcovn56/'
             ],
             'INVOICE'                      => [
                 's-ivc-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-ivc-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/invoice/s-ivc-xen2ybcovn56/'
             ],
             'INVOICE_GUARANTEED'           => [
                 's-ivg-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-ivg-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/invoice-guaranteed/s-ivg-xen2ybcovn56/'
             ],
             'PAYPAL'                       => [
                 's-ppl-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-ppl-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/paypal/s-ppl-xen2ybcovn56/'
             ],
             'PREPAYMENT'                   => [
                 's-ppy-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-ppy-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/prepayment/s-ppy-xen2ybcovn56/'
             ],
             'PRZELEWY24'                   => [
                 's-p24-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-p24-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/przelewy24/s-p24-xen2ybcovn56/'
             ],
             'SEPA_DIRECT_DEBIT_GUARANTEED' => [
                 's-ddg-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-ddg-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/direct-debit-guaranteed/s-ddg-xen2ybcovn56/'
             ],
             'SEPA_DIRECT_DEBIT'            => [
                 's-sdd-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-sdd-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/direct-debit/s-sdd-xen2ybcovn56/'
             ],
             'SOFORT'                       => [
                 's-sft-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-sft-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/sofort/s-sft-xen2ybcovn56/'
             ],
             'PIS'                          => [
                 's-pis-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-pis-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/pis/s-pis-xen2ybcovn56/'
             ],
             'EPS'                          => [
                 's-eps-xen2ybcovn56',
-                'https://api.heidelpay.com/v1/types/card/s-eps-xen2ybcovn56/'
+                'https://api.heidelpay.com/v1/types/eps/s-eps-xen2ybcovn56/'
+            ],
+            'ALIPAY'                       => [
+                's-ali-xen2ybcovn56',
+                'https://api.heidelpay.com/v1/types/alipay/s-ali-xen2ybcovn56/'
+            ],
+            'WECHATPAY'                    => [
+                's-wcp-xen2ybcovn56',
+                'https://api.heidelpay.com/v1/types/wechatpay/s-wcp-xen2ybcovn56/'
+            ],
+            'INVOICE_FACTORING'            => [
+                's-ivf-xen2ybcovn56',
+                'https://api.heidelpay.com/v1/types/wechatpay/s-ivf-xen2ybcovn56/'
             ]
         ];
     }